refactor: Extract useVisible
